+++ src/Updater/UpdateThemes.php
@@ -26,11 +26,6 @@ class UpdateThemes
         $this->cache_key = UNREPRESS_PREFIX . 'updates_theme_';
         $this->cache_results = true;
 
-        // If force-check=1 and page=unrepress-updates, then empty all update transients
-        if (isset($_GET['force-check']) && $_GET['force-check'] === '1' && isset($_GET['page']) && $_GET['page'] === 'unrepress-updater') {
-            $this->deleteAllUpdateTransients();
-        }
-
         $this->checkforUpdates();
 
         add_filter('themes_api', [$this, 'getInformation'], 20, 3);
@@ -73,9 +68,9 @@ class UpdateThemes
                 $theme = wp_get_theme($slug);
 
                 // Get data from the remote source
-                $updateInfo->requires = $remoteData->requires ?? '';
-                $updateInfo->tested = $remoteData->tested ?? '';
-                $updateInfo->requires_php = $remoteData->requires_php ?? '';
+                $updateInfo->requires = $remoteData->requires ?? '6.5';
+                $updateInfo->tested = $remoteData->tested ?? '6.7';
+                $updateInfo->requires_php = $remoteData->requires_php ?? '8.1';
 
                 // Get data from the local theme object
                 $updateInfo->name = $theme->get('Name');
@@ -341,48 +336,16 @@ class UpdateThemes
     }
 
     /**
-     * Deletes all transients used by the updates API.
-     *
-     * All transients used by the updates API have a name that begins with
-     * UNREPRESS_PREFIX . 'updates_theme'. This method will delete all
-     * transients with names that match this pattern.
-     *
-     * @since 1.0.0
-     */
-    private function deleteAllUpdateTransients()
-    {
-        global $wpdb;
-
-        // Delete both transients and their timeout entries
-        $wpdb->query(
-            $wpdb->prepare(
-                "DELETE FROM {$wpdb->options}
-                WHERE option_name LIKE %s
-                OR option_name LIKE %s",
-                '_transient_' . $this->cache_key . '%',
-                '_transient_timeout_' . $this->cache_key . '%'
-            )
-        );
-    }
-
-    /**
      * Fix source directory for GitHub theme updates
      */
     public function maybeFixSourceDir($source, $remote_source, $upgrader, $args)
     {
         if (! isset($args['theme'])) {
-            Debugger::log("Theme not found in args");
             return $source;
         }
 
-        Debugger::log("Fixing theme source directory");
-        Debugger::log("Source: {$source}");
-        Debugger::log("Remote source: {$remote_source}");
-        Debugger::log("Theme: {$args['theme']}");
-
         $result = Helpers::fixSourceDir($source, $remote_source, $args['theme'], 'theme');
 
-        Debugger::log("New source directory: {$result}");
         return $result;
     }
 }