+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # UnrePress changelog
 
-## 0.2.0 - 2024-12-24
+## 0.3.0 - 2024-12-24
 * Working themes updater.
+* Working plugins updater.
+* Tweaked core updater.
+* Spread the word!
 
-## 0.1.4 - 2024-12-23
+## 0.2.0 - 2024-12-23
 * Working plugins updater.
 
 ## 0.1.0 - 2024-11-03